feat(dashboard): move BYOK UI from /credentials to /settings/byok (RFC PR 6)
PR 6 of the Queen Execution Mode RFC's 6-PR stack. Independent of
PRs 1-5; can land any time. Purely cosmetic — same components,
same Redis storage key, same operator session auth.

What changes:
- web/src/app/dashboard/credentials/{page,CredentialsPanel}.tsx
  → web/src/app/dashboard/settings/byok/{page,ByokPanel}.tsx
  (git rename preserved so blame survives the move)
- /dashboard/credentials becomes a single-line redirect stub
  pointing at /dashboard/settings/byok. Stale bookmarks keep
  working.
- DashboardNav points at the new path; label stays "Credentials"
  so the operator UX text is unchanged.

What does NOT change:
- hive:byok:{installationId} Redis key — explicitly out of scope
  per the RFC. The key was identified as load-bearing across
  multiple call sites (web BYOK store, bot llm/byok.ts,
  byok-auth.ts caching) and any rename would need a separate
  migration plan with key-rotation handling.
- Component contents. CredentialsPanel renames to ByokPanel for
  symmetry with PR 5's upcoming sibling at /dashboard/settings,
  but no logic changes.

Tests: 1 redirect-target unit test + full web suite (1480 passing,
no regressions).

Refs RFC PR 6 row.
